+++ react/typings/hoist-non-react-statics.d.ts
@@ -1,0 +1,5 @@
+declare module 'hoist-non-react-statics' {
+  const hoistNonReactStatic: any
+
+  export default hoistNonReactStatic
+}