+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `hoist-non-react-statics` and support to static `propTypes` to `withCssHandles`.
 
 ## [0.4.0] - 2019-10-08
 ### Added